requirements.txt
ultralytics>=8.0.0
pandas>=1.5.0
numpy>=1.23.0
Pillow>=9.0.0
matplotlib>=3.6.0
scikit-learn>=1.1.0
PyYAML>=6.0
tqdm>=4.64.0
pytest>=7.2.0
